requirements.txt
# --- Основные библиотеки ---
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.4.2

# --- Работа с LLM (Ollama) ---
requests==2.32.3

# --- Эмбеддинги и RAG ---
sentence-transformers==2.7.0
faiss-cpu==1.8.0
nltk==3.8.1

# --- Проверка качества (RAGChecker) ---
ragas==0.1.7
evaluate==0.4.2

# --- Torch (CPU версия для совместимости) ---
torch==2.2.2
torchvision==0.17.2
torchaudio==2.2.2

# --- Утилиты ---
tqdm==4.66.4
python-dotenv==1.0.1

# --- JSON и файлы ---
orjson==3.10.3

# --- Дополнительно для стабильности ---
typing_extensions==4.12.2
